requirements.txt
ffmpeg_python==0.2.0
matplotlib==3.9.2
numpy==2.1.3
opencv_python==4.10.0.84
